Moved classes to package private within the MicrosoftTeamsNotifier. Moved all classes under the de.codecentric.boot.admin.notify package. Assert the specific values of the message generation methods to give specific errors when something breaks. Also refactored to just a single postForObject call expecting the Void.class. Also made the components of a message non-final so they could be overridden by application properties/yaml file. If there are an incorrect number of formatting arguments, will just use the message from the application properties/yaml as is.
