+++ spring-boot-admin-server/src/main/java/de/codecentric/boot/admin/config/NotifierConfiguration.java
@@ -19,6 +19,7 @@ import de.codecentric.boot.admin.notify.CompositeNotifier;
 import de.codecentric.boot.admin.notify.HipchatNotifier;
 import de.codecentric.boot.admin.notify.LetsChatNotifier;
 import de.codecentric.boot.admin.notify.MailNotifier;
+import de.codecentric.boot.admin.notify.MicrosoftTeamsNotifier;
 import de.codecentric.boot.admin.notify.Notifier;
 import de.codecentric.boot.admin.notify.NotifierListener;
 import de.codecentric.boot.admin.notify.OpsGenieNotifier;
@@ -26,7 +27,6 @@ import de.codecentric.boot.admin.notify.PagerdutyNotifier;
 import de.codecentric.boot.admin.notify.SlackNotifier;
 import de.codecentric.boot.admin.notify.filter.FilteringNotifier;
 import de.codecentric.boot.admin.notify.filter.web.NotificationFilterController;
-import de.codecentric.boot.admin.notify.microsoft.teams.MicrosoftTeamsNotifier;
 import de.codecentric.boot.admin.web.PrefixHandlerMapping;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
